requirements.txt
pytest>=8.0
ruff>=0.5
mypy>=1.10
PyYAML>=6.0
types-PyYAML>=6.0.12.20240808
pre-commit>=3.7
